+++ composeApp/src/commonMain/kotlin/App.kt
@@ -1,7 +1,7 @@
 import androidx.compose.runtime.Composable
 import cafe.adriel.voyager.navigator.Navigator
-import core.theme.ui.ThemeProvider
-import features.home.ui.screens.home.HomeScreen
+import core.presentation.theme.ThemeProvider
+import features.home.presentation.screens.home.HomeScreen
 
 @Composable
 fun App() {
