+++ composeApp/src/commonMain/kotlin/core/presentation/composables/error/Error.kt
@@ -22,7 +22,7 @@ import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import coil3.compose.AsyncImage
-import core.theme.ui.Theme
+import core.presentation.theme.Theme
 import org.jetbrains.compose.ui.tooling.preview.Preview
 
 @Composable
