+++ composeApp/src/commonMain/kotlin/core/presentation/theme/Theme.kt
@@ -1,4 +1,4 @@
-package core.theme.ui
+package core.presentation.theme
 
 import androidx.compose.material.Colors
 import androidx.compose.material.MaterialTheme
