+++ composeApp/src/commonMain/kotlin/core/presentation/theme/ThemeProvider.kt
@@ -1,4 +1,4 @@
-package core.theme.ui
+package core.presentation.theme
 
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material.MaterialTheme
