+++ composeApp/src/commonMain/kotlin/core/presentation/theme/Typography.kt
@@ -1,4 +1,4 @@
-package core.theme.ui
+package core.presentation.theme
 
 import androidx.compose.material.Typography
 import androidx.compose.runtime.Composable
