+++ composeApp/src/commonMain/kotlin/features/character/di/CharacterModule.kt
@@ -1,6 +1,6 @@
 package features.character.di
 
-import features.home.ui.screens.home.HomeScreenModel
+import features.home.presentation.screens.home.HomeScreenModel
 import org.koin.dsl.module
 
 val characterModule = module {
