+++ composeApp/src/commonMain/kotlin/features/character/presentation/screens/characterDetail/CharacterDetailScreenModel.kt
@@ -1,4 +1,4 @@
-package features.character.ui.screens.characterDetail
+package features.character.presentation.screens.characterDetail
 
 import cafe.adriel.voyager.core.model.StateScreenModel
