+++ composeApp/src/commonMain/kotlin/features/character/presentation/screens/characterDetail/CharacterDetailState.kt
@@ -1,4 +1,4 @@
-package features.character.ui.screens.characterDetail
+package features.character.presentation.screens.characterDetail
 
 data class CharacterDetailState(
     val testParam: Boolean = false
