+++ composeApp/src/commonMain/kotlin/features/home/di/HomeModule.kt
@@ -1,6 +1,6 @@
 package features.home.di
 
-import features.character.ui.screens.characterDetail.CharacterDetailScreenModel
+import features.character.presentation.screens.characterDetail.CharacterDetailScreenModel
 import org.koin.dsl.module
 
 val homeModule = module {
